third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-script-cors-no-xorigin-console-messages-expected.txt

CONSOLE ERROR: Subresource Integrity: The resource 'http://localhost:8000/security/resources/localScript.js' has an integrity attribute, but the resource requires the request to be CORS enabled to check the integrity, and it is not. The resource has not been blocked, but no integrity check occurred.
